Combine test files
